Top birders list
This is the  for entering TOP EDGE scores in a persistent way. It has the flaw of not being automated, the score has to be reentered manually. It wont work on shiny apps. Only in your own computer or if you have access to the server where the app resides.